feat: new `createDelayHoverSignal`
